requirements.txt
amqp==5.2.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
bootstrap4==0.1.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Django==5.0.2
django-bootstrap4==24.1
django-cron==0.6.0
django-crontab==0.7.1
django-preventconcurrentlogins==0.8.2
django-simple-history==3.5.0
django-tenants==3.6.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
fontawesomefree==6.0.0
fonttools==4.25.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kombu==5.3.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
munkres==1.1.4
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
numpy==1.26.4
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.9.0
Pygments==2.17.2
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
yarg==0.1.9
zopfli==0.2.3
aiohttp==3.9.3
azure-identity==1.15.0
httpx==0.27.0
msal==1.28.0
msgraph-core==1.0.0
python-dotenv==1.0.1
requests==2.31.0